requirements/requirements.txt
# Local Segmentation
-e .

# ProtoSeg
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
boto3==1.21.8
botocore==1.24.8
bravado==11.0.3
bravado-core==5.17.0
cachetools==5.0.0
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
cython==0.29.36 # --install-option="--force-reinstall" Potentially needed
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.2
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
gin-config==0.5.0
gitdb==4.0.9
gitpython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.2
ipykernel==6.9.1
ipython==8.1.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isoduration==20.11.0
jedi==0.18.1
jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.2
jsonref==0.2
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
markdown==3.3.6
markupsafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.6
msgpack==1.0.3
multidict==6.0.2
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
oauthlib==3.2.0
omegaconf==2.2.2
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeprecate==0.3.1
pygments==2.11.2
pyjwt==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2021.3
pyyaml==6.0
pyzmq==22.3.0
qtconsole==5.2.2
qtpy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
rsa==4.8
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
send2trash==1.8.0
setuptools==59.5.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0
smmap==5.0.0
stack-data==0.2.0
swagger-spec-validator==2.7.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.1.0
torch==1.10.2
torchfile==0.1.0
torchmetrics==0.7.2
torchnet==0.0.4
torchvision==0.11.3
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
transformers==4.34.1
typing-extensions==4.1.1
uri-template==1.1.0
urllib3==1.26.8
visdom==0.1.8.9
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.3.1
werkzeug==2.0.3
widgetsnbextension==3.5.2
yarl==1.7.2
zipp==3.7.0

# Extended Requirements
python-dotenv
wandb
panoptic-parts
